♻️ moved dependencies from dev to normal moved schemas to src directory
