+++ src/back-end/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "private":"true",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "nodemon src/app.ts"
@@ -20,7 +21,18 @@
     "ejs": "^3.1.6",
     "mogoose": "0.0.1-security",
     "pino-pretty": "^5.1.3",
-    "socket.io": "^4.1.3"
+    "socket.io": "^4.1.3",
+    "bcrypt": "^5.0.1",
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
+    "jsonwebtoken": "^8.5.1",
+    "lodash": "^4.17.21",
+    "mongoose": "^5.13.5",
+    "nanoid": "^3.1.23",
+    "pino": "^6.13.0",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5",
+    "yup": "^0.32.9"
   },
   "devDependencies": {
     "@types/bcrypt": "^5.0.0",
@@ -33,16 +45,5 @@
     "@types/node": "^16.4.13",
     "@types/pino": "^6.3.11",
     "@types/yup": "^0.29.13",
-    "bcrypt": "^5.0.1",
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
-    "lodash": "^4.17.21",
-    "mongoose": "^5.13.5",
-    "nanoid": "^3.1.23",
-    "pino": "^6.13.0",
-    "ts-node": "^10.1.0",
-    "typescript": "^4.3.5",
-    "yup": "^0.32.9"
   }
 }
